Retrieval/requirements.txt
# ── LLM Providers ──────────────────────────────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-text-splitters>=0.3.0
langchain-experimental>=0.3.0
langgraph>=0.2.0

# ── LLM Clients ────────────────────────────────────────────
openai>=1.50.0
anthropic>=0.36.0

# ── Vector Store ────────────────────────────────────────────
faiss-cpu>=1.8.0

# ── Embeddings ──────────────────────────────────────────────
sentence-transformers>=3.0.0

# ── Knowledge Graph ─────────────────────────────────────────
neo4j>=5.24.0

# ── Document Loaders ────────────────────────────────────────
pypdf>=4.0.0
python-docx>=1.1.0
unstructured>=0.15.0
unstructured[html]>=0.15.0

# ── Caching ──────────────────────────────────────────────────
redis>=5.1.0

# ── API ─────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.12

# ── Config ──────────────────────────────────────────────────
pydantic>=2.9.0
pydantic-settings>=2.5.0

# ── Evaluation ──────────────────────────────────────────────
ragas>=0.2.0
datasets>=3.0.0

# ── UI ───────────────────────────────────────────────────────
streamlit>=1.39.0

# ── Utilities ────────────────────────────────────────────────
numpy>=1.26.0
pandas>=2.2.0
python-dotenv>=1.0.0
httpx>=0.27.0

# ── Testing ──────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.24.0
